refactored drive off tarmac command to drive distance command fixed check style errors
